+++ simulator/messages/movevesseloutputmessage.h
@@ -17,7 +17,7 @@ public:
     }
 
     bool send(std::ostream &strm);
-    int sendBinary(void *buffer, int maxlen);
+    int sendBinary(void *buffer, size_t maxlen);
 
 private:
     unsigned int tstep, tstep_dep;
